requirement.txt
appdirs==1.4.4
APScheduler==3.10.4
argcomplete==3.1.2
blinker==1.6.3
click==8.1.7
escpos==2.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
future==0.18.3
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
Pillow==10.1.0
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
pytz==2023.3.post1
pyusb==1.2.1
PyYAML==6.0.1
qrcode==7.4.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzlocal==5.2
viivakoodi==0.8.0
Werkzeug==3.0.1